% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppRegpanelmixMLE.R
\name{cppregpanelmixMLE}
\alias{cppregpanelmixMLE}
\title{cppregpanelmixMLE}
\usage{
cppregpanelmixMLE(ys, xs, zs, alpha0s, mu0s, sigma0s, beta0s, gamma0s, m, q, p,
  t)
}
\arguments{
\item{m}{The number of components in the mixture}

\item{q}{The number of variable in x}

\item{p}{The number of variable in z}

\item{alpha}{m by 1 vector}

\item{mu}{m by 1 vector}

\item{sigma}{m by 1 vector}

\item{gamma}{p by 1 vector}

\item{beta}{m by q matrix}

\item{y}{nt by 1 vector of data}

\item{x}{nt by q matrix of data for x (if exists)}

\item{z}{nt by p matrix of regressor associated with gamma}
}
\value{
log-likelihood
}
\description{
return panel data likelihood
}

