% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoefficient.R
\name{getCoefficient}
\alias{getCoefficient}
\title{get coefficients name}
\usage{
getCoefficient(coefficients, N, T, M, p, q)
}
\arguments{
\item{N}{int}

\item{T}{int}

\item{M}{number of types}

\item{p}{columns of Z}

\item{q}{columns of X}
}
\value{
list of item
\item{alpha}
\item{mu}
\item{beta}
\item{sigma}
\item{gamma}
}
\description{
get a list of coeficients from output
}

