\name{normalRegPanelMix-package}
\alias{normalRegPanelMix-package}
\alias{normalRegPanelMix}
\docType{package}
\title{
\packageTitle{normalRegPanelMix}
}
\description{
\packageDescription{normalRegPanelMix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{normalRegPanelMix}
\packageIndices{normalRegPanelMix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{normalRegPanelMix}

Maintainer: \packageMaintainer{normalRegPanelMix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
