% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regpanelmix_crit.R
\name{regpanelmixCrit}
\alias{regpanelmixCrit}
\title{regpanelmixCrit}
\usage{
regpanelmixCrit(y, x, parlist, z = NULL, values = NULL, parallel = TRUE,
  cl = NULL, nrep = 1000, ninits.crit = 25, deep.check = 0)
}
\arguments{
\item{parallel}{Determines whether package \code{doParallel} is used for calculation}

\item{cl}{Cluster used for parallelization; if it is \code{NULL}, the system will automatically}

\item{nrep}{The number of replications used to compute p-values}

\item{ninits.crit}{The number of initial guesses to form critical values}
}
\value{
A list with the following items:
\item{crit}{3 by 3 matrix of (0.1, 0.05, 0.01 critical values), jth row corresponding to k=j}
\item{pvals}{A vector of p-values at k = 1, 2, 3}
}

