% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNormalpanelmixMLE.R
\name{rnormalpanelmixlikelihood}
\alias{rNormalPanelMixMLE}
\alias{rnormalpanelmixlikelihood}
\title{rnormalpanelmixlikelihood}
\usage{
rNormalPanelMixMLE(alpha, mu, sigma, beta, gamma, y, x, z, M, p, q)
}
\arguments{
\item{y}{nt by 1 vector of data}

\item{x}{nt by q matrix of data for x (if exists)}

\item{z}{nt by p matrix of regressor associated with gamma}

\item{p}{The number of variable in z}

\item{q}{The number of variable in x}

\item{b}{(p+q+3)*m by 1 vector, b <- (alpha,mu,beta,sigma,q)}

\item{m}{The number of components in the mixture}
}
\value{
log-likelihood
}
\description{
Estimates parameters of a finite panel mixture of univariate normals by
penalized maximum log-likelhood functions.
}

