% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNormalpanelmixMLE_vec.R
\name{rNormalPanelMixMLE_vec}
\alias{rNormalPanelMixMLE_vec}
\title{rNormalPanelMixMLE_vec}
\usage{
rNormalPanelMixMLE_vec(b0, y, x, z, M, p, q)
}
\arguments{
\item{y}{nt by 1 vector of data}

\item{x}{nt by q matrix of data for x (if exists)}

\item{z}{nt by p matrix of regressor associated with gamma}

\item{p}{The number of variable in z}

\item{q}{The number of variable in x}

\item{alpha}{m by 1 vector}

\item{mu}{m by 1 vector}

\item{sigma}{m by 1 vector}

\item{gamma}{p by 1 vector}

\item{beta}{m by q matrix}

\item{m}{The number of components in the mixture}
}
\value{
log-likelihood
}
\description{
Estimates parameters of a finite panel mixture of univariate normals by
penalized maximum log-likelhood functions.
}

