% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funcs.R
\name{rnormregpanelmix}
\alias{rnormregpanelmix}
\title{rnormregpanelmix}
\usage{
rnormregpanelmix(n, t, x = NULL, alpha, mubeta, sigma)
}
\arguments{
\item{n}{The number of observations}

\item{t}{Number of time periods}

\item{x}{n by k-1 matrix that does NOT include a constant}

\item{alpha}{m by 1 vector that represents proportions of components}

\item{mubeta}{k by m matrix that represents (mu times k regression coefficients) on x for m components}

\item{sigma}{m by 1 vector that represents sd of components}
}
\value{
n by 1 vector that is formed by regressor x
}
\description{
Generates mixed normal random variables with regressor x
}

